% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicsApp-package.R
\docType{package}
\name{omicsApp-package}
\alias{omicsApp}
\alias{omicsApp-package}
\title{omicsApp: Shiny Web Interface for Multi-Omics Analysis}
\description{
Interactive Shiny application for proteomics and transcriptomics analysis,
powered by the \href{https://github.com/puweilin/omicsApp}{omicsCore} engine.
}
\details{
The entry point is \code{\link[=launch]{launch()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/puweilin/omicsApp}
  \item Report bugs at \url{https://github.com/puweilin/omicsApp/issues}
}

}
\author{
\strong{Maintainer}: Weilin Pu \email{puweilin@gmail.com}

}
\keyword{internal}
